+++ Runtime/Core/Consoles/Console.cs
@@ -3,38 +3,38 @@ using System.Globalization;
 using System.Text;
 using UnityEngine;
 
-namespace BaseTool.Core.Consoles
+namespace BaseTool
 {
     public class Console
     {
         private const int HistoryCount = 50;
-        
+
         private static string LastMsg = "";
         private static double TimeLastMsg;
         private static int ConsoleShowLastLine = 0;
-        
+
         private static List<string> PendingCommands = new();
         public static int PendingCommandsWaitForFrames = 0;
         public static bool PendingCommandsWaitForLoad = false;
-        
-        private static Dictionary<string, ConsoleCommand> _commands = new();    
+
+        private static Dictionary<string, ConsoleCommand> _commands = new();
         private static string[] History = new string[HistoryCount];
         private static int HistoryNextIndex = 0;
         private static int HistoryIndex = 0;
-        
+
         private static ConsoleManager _consoleManager;
 
         public Console()
         {
             var go = new GameObject("Console (manager)");
             Object.DontDestroyOnLoad(go);
-            
+
             _consoleManager = go.AddComponent<ConsoleManager>();
-            
+
             AddCommand("help", "How to use the command", HelpCommand);
             AddCommand("list", "Display the list of commands", ListCommand);
-        }   
-        
+        }
+
         private static void OutputString(string message)
         {
             if (_consoleManager == null)
@@ -51,7 +51,7 @@ namespace BaseTool.Core.Consoles
             }
             OutputString(msg);
         }
-        
+
         private static string ParseQuoted(string input, ref int pos)
         {
             ++pos;
@@ -78,8 +78,8 @@ namespace BaseTool.Core.Consoles
                 ++pos;
             }
             return input.Substring(startIndex);
-        }    
-        
+        }
+
         private static List<string> Tokenize(string input)
         {
             int pos = 0;
@@ -118,20 +118,20 @@ namespace BaseTool.Core.Consoles
             ++HistoryIndex;
             return History[HistoryIndex % 50];
         }
-        
+
         public static ConsoleArguments ParseArguments(string[] argument) => new ConsoleArguments(argument);
-        
+
         private static void SkipWhite(string input, ref int pos)
         {
             while (pos < input.Length && " \t".IndexOf(input[pos]) > -1)
                 ++pos;
         }
-        
+
         #region UPDATES
-        
+
         public static void ConsoleUpdate()
         {
-            double num = (double) Time.time - TimeLastMsg;
+            double num = (double)Time.time - TimeLastMsg;
             //IconikFramework.Console.Console.s_ConsoleUI.ConsoleUpdate();
             while (PendingCommands.Count > 0)
             {
@@ -147,11 +147,11 @@ namespace BaseTool.Core.Consoles
                 ExecuteCommand(pendingCommand);
             }
         }
-        
+
         #endregion
-        
+
         #region COMMAND EXECUTION
-        
+
         public static void ExecuteCommand(string command)
         {
             List<string> stringList = Tokenize(command);
@@ -177,14 +177,14 @@ namespace BaseTool.Core.Consoles
             HistoryIndex = HistoryNextIndex;
             EnqueueCommandNoHistory(command);
         }
-        
+
         public static string TabComplete(string prefix)
         {
             List<string> stringList = new List<string>();
             foreach (KeyValuePair<string, ConsoleCommand> command in _commands)
             {
                 string key = command.Key;
-                if (key.StartsWith(prefix, true, (CultureInfo) null))
+                if (key.StartsWith(prefix, true, (CultureInfo)null))
                     stringList.Add(key);
             }
             if (stringList.Count == 0)
@@ -208,16 +208,16 @@ namespace BaseTool.Core.Consoles
             int num = Mathf.Min(a.Length, b.Length);
             for (int length = 1; length <= num; ++length)
             {
-                if (!a.StartsWith(b.Substring(0, length), true, (CultureInfo) null))
+                if (!a.StartsWith(b.Substring(0, length), true, (CultureInfo)null))
                     return length - 1;
             }
             return num;
         }
-        
+
         #endregion
-        
+
         #region COMMANDS
-        
+
         public static void AddCommand(
             string name,
             string description,
@@ -247,22 +247,22 @@ namespace BaseTool.Core.Consoles
             }
             OutputString(sb.ToString());
         }
-        
+
         #endregion
-        
+
         #region AUTOLOAD
 
         private static Console _instance;
 
         public static Console Instance => _instance ?? Init();
-        
+
         [RuntimeInitializeOnLoadMethod]
         public static Console Init()
         {
             _instance = new Console();
             return _instance;
         }
-        
+
         #endregion
     }
 }