+++ Runtime/Core/Tools/Attributes/IfNotAttribute.cs
@@ -1,0 +1,10 @@
+using System;
+
+namespace BaseTool
+{
+    [AttributeUsage(AttributeTargets.Field)]
+    public class IfNotAttribute : ConditionalAttribute
+    {
+        public IfNotAttribute(string selector) : base(selector) { }
+    }
+}