+++ Runtime/Movement/FirstPersonController.cs
@@ -6,42 +6,48 @@ namespace BaseTool.Movement
     [RequireComponent(typeof(Rigidbody))]
     public class FirstPersonController : MonoBehaviour, IMovable
     {
-        [GetComponent]
-        private Rigidbody _rigidbody;
+        [GetComponent, SerializeField]
+        protected Rigidbody _rigidbody;
 
-        [GetComponentInChildren]
-        private Camera _camera;
+        [GetComponentInChildren, SerializeField]
+        protected Camera _camera;
 
         [Header("Movement Settings")]
         [SerializeField]
-        private float _speed = 5;
+        protected float _speed = 5;
 
         [SerializeField]
-        private float _lookSpeed = 5;
+        protected float _lookSpeed = 5;
 
-        private Vector2 _moveInput = Vector2.zero;
-        private Vector2 _rotationInput = Vector2.zero;
+        protected Vector2 _moveInput = Vector2.zero;
+        protected Vector2 _rotationInput = Vector2.zero;
 
-        private float _cameraRotation = 0;
+        protected float _cameraRotation = 0;
 
         protected virtual void Awake() => Injector.Process(this);
 
         protected virtual void FixedUpdate() => FixedUpdateMove();
 
-        protected void FixedUpdateMove()
+        protected virtual void FixedUpdateMove()
         {
             Vector3 tmp = new Vector3(_moveInput.x, 0, _moveInput.y);
-            var dir = _speed * transform.TransformDirection(tmp);
+            var dir = _speed * transform.TransformDirection(tmp.LimitLength());
+
+#if UNITY_2023_3_OR_NEWER
+            dir.y = _rigidbody.linearVelocity.y;
+            _rigidbody.linearVelocity = dir;
+#else
             dir.y = _rigidbody.velocity.y;
             _rigidbody.velocity = dir;
+#endif
         }
 
-        public void Move(Vector2 move)
+        public virtual void Move(Vector2 move)
         {
             _moveInput = move;
         }
 
-        public void Rotate(Vector2 rotation)
+        public virtual void Rotate(Vector2 rotation)
         {
             _rotationInput = rotation;
             var newRot = _rigidbody.rotation * Quaternion.Euler(_lookSpeed * Time.fixedDeltaTime * _rotationInput.x * Vector3.up);