+++ Runtime/Movement/SideViewController.cs
@@ -14,9 +14,15 @@ namespace BaseTool.Movement
 
         public void Move(Vector2 move)
         {
+#if UNITY_2023_3_OR_NEWER
+            var velocity = _rigidbody.linearVelocity;
+            velocity.x = _moveSpeed * move.x;
+            _rigidbody.linearVelocity = velocity;
+#else
             var velocity = _rigidbody.velocity;
             velocity.x = _moveSpeed * move.x;
             _rigidbody.velocity = velocity;
+#endif
         }
 
         public void Rotate(Vector2 _) { }