+++ Runtime/Shooter/ContactData.cs
@@ -1,0 +1,11 @@
+﻿using UnityEngine;
+
+namespace BaseTool.Shooter
+{
+    public struct ContactData
+    {
+        public GameObject Object;
+        public Vector3 HitPoint;
+        public Vector3 Normal;
+    }
+}