+++ Runtime/Shooter/Interfaces/IShootController.cs
@@ -1,0 +1,13 @@
+﻿using System;
+
+namespace BaseTool.Shooter
+{
+    public interface IShootController
+    {
+        public event Action OnStartShoot;
+
+        public event Action OnStopShoot;
+
+        public event Action OnReload;
+    }
+}