+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Improvements
 
+- Package documentation enhanced
 - Add the `SetupWizard` to enable or disable modules [[#7](https://github.com/DarkRewar/BaseTool/pull/7)]
 - Add scriptable object `GameEvent` system  [[#11](https://github.com/DarkRewar/BaseTool/pull/11)]